working parser

+++ build_parser.py
@@ -7,8 +7,8 @@ PARSER_DIR = "compiler0/generated"
 if os.path.exists(PARSER_DIR):
     # remove the generated parser
     for file in os.listdir(PARSER_DIR):
-        # if file is not __init__.py
-        if file != "__init__.py":
+        # if is a python file and not __init__.py
+        if file.endswith(".py") and file != "__init__.py":
             os.remove(os.path.join(PARSER_DIR, file))
 
 call(["antlr4", "-Dlanguage=Python3", "-visitor", "-o", PARSER_DIR, "Quanta.g4"])
